test: consolidate coverage suites
Why:
- The coverage campaign scattered related tests across 18 narrowly named coverage files, making package test organization harder to navigate.

What:
- Fold backend, compiler, GC, and public API tests into existing behavior-focused suites.
- Rename the remaining corpus and validator suites to describe their behavior rather than the coverage mechanism.
- Preserve every test/benchmark declaration and all platform build constraints.

Proof:
- Test files: 325 -> 307
- Test/benchmark declaration inventory unchanged
- make test
- Go 1.22 linux/amd64 test binaries compile
- Coverage preserved: arm64 89.9%, IR 88.5%, wasm 87.0%, GC 88.5%; wago improves 80.3% -> 83.1%

Docs:
- unchanged, test organization only with no workflow impact.
